// src/printer-model/types.ts
export type PrusaPrinterModelGLTF = GLTF & {
  nodes: {
    frame_long_extrusion_1: THREE.Mesh;
    frame_long_extrusion_2: THREE.Mesh;
    frame_mk3_front_plate: THREE.Mesh;
    frame_mk3_rear_plate: THREE.Mesh;
    frame_mk3s: THREE.Mesh;
    frame_screw_1: THREE.Mesh;
    frame_screw_10: THREE.Mesh;
    frame_screw_11: THREE.Mesh;
    frame_screw_12: THREE.Mesh;
    frame_screw_13: THREE.Mesh;
    frame_screw_14: THREE.Mesh;
    frame_screw_15: THREE.Mesh;
    frame_screw_16: THREE.Mesh;
    frame_screw_17: THREE.Mesh;
    frame_screw_18: THREE.Mesh;
    frame_screw_19: THREE.Mesh;
    frame_screw_2: THREE.Mesh;
    frame_screw_20: THREE.Mesh;
    frame_screw_3: THREE.Mesh;
    frame_screw_4: THREE.Mesh;
    frame_screw_5: THREE.Mesh;
    frame_screw_6: THREE.Mesh;
    frame_screw_7: THREE.Mesh;
    frame_screw_8: THREE.Mesh;
    frame_screw_9: THREE.Mesh;
    frame_short_extrusion_1: THREE.Mesh;
    frame_short_extrusion_2: THREE.Mesh;
    frog_base: THREE.Mesh;
    frog_belt_holder: THREE.Mesh;
    frog_belt_tensioner: THREE.Mesh;
    frog_u_bolt_1: THREE.Mesh;
    frog_u_bolt_2: THREE.Mesh;
    frog_u_bolt_3: THREE.Mesh;
    heatbed_base_1: THREE.Mesh;
    heatbed_base_2: THREE.Mesh;
    heatbed_base_sheet_1: THREE.Mesh;
    heatbed_base_sheet_2: THREE.Mesh;
    heatbed_cable_cover_bottom: THREE.Mesh;
    heatbed_cable_cover_top: THREE.Mesh;
    heatbed_space_1: THREE.Mesh;
    heatbed_space_2: THREE.Mesh;
    heatbed_space_3: THREE.Mesh;
    heatbed_space_4: THREE.Mesh;
    heatbed_space_5: THREE.Mesh;
    heatbed_space_6: THREE.Mesh;
    heatbed_space_7: THREE.Mesh;
    heatbed_space_8: THREE.Mesh;
    heatbed_space_9: THREE.Mesh;
    lcd_cover: THREE.Mesh;
    lcd_supports_1: THREE.Mesh;
    lcd_supports_2: THREE.Mesh;
    lcd_board_1: THREE.Mesh;
    lcd_board_2: THREE.Mesh;
    lcd_board_3: THREE.Mesh;
    lcd_board_4: THREE.Mesh;
    lcd_knob: THREE.Mesh;
    extruder_cooling_fan_1: THREE.Mesh;
    extruder_cooling_fan_2: THREE.Mesh;
    extruder_cooling_fan_3: THREE.Mesh;
    extruder_noctua_1: THREE.Mesh;
    extruder_noctua_2: THREE.Mesh;
    extruder_noctua_3: THREE.Mesh;
    extruder_noctua_4: THREE.Mesh;
    extruder_noctua_5: THREE.Mesh;
    extruder_nema_1: THREE.Mesh;
    extruder_nema_2_1: THREE.Mesh;
    extruder_nema_2_2: THREE.Mesh;
    extruder_nema_10_1: THREE.Mesh;
    extruder_nema_10_2: THREE.Mesh;
    extruder_nema_10_3: THREE.Mesh;
    extruder_nema_10_4: THREE.Mesh;
    extruder_nema_3: THREE.Mesh;
    extruder_nema_4: THREE.Mesh;
    extruder_nema_5: THREE.Mesh;
    extruder_nema_7: THREE.Mesh;
    extruder_nema_8: THREE.Mesh;
    extruder_nema_9: THREE.Mesh;
    extruder_x_carriage_back_2: THREE.Mesh;
    extruder_body: THREE.Mesh;
    extruder_cable_holder: THREE.Mesh;
    extruder_cover: THREE.Mesh;
    extruder_e3d: THREE.Mesh;
    extruder_fan_shroud: THREE.Mesh;
    extruder_fs_cover: THREE.Mesh;
    extruder_idler: THREE.Mesh;
    extruder_motor_plate: THREE.Mesh;
    extruder_pinda_probe_1: THREE.Mesh;
    extruder_pinda_probe_2: THREE.Mesh;
    extruder_print_fan_support: THREE.Mesh;
    extruder_x_carriage_back: THREE.Mesh;
    y_nema_1: THREE.Mesh;
    x_nema_1: THREE.Mesh;
    x_nema_3: THREE.Mesh;
    psu_plug_1: THREE.Mesh;
    psu_plug_2: THREE.Mesh;
    psu_switch001: THREE.Mesh;
    psu_cover_bottom: THREE.Mesh;
    psu_cover_top: THREE.Mesh;
    psu_metal: THREE.Mesh;
    einsy_base: THREE.Mesh;
    einsy_door: THREE.Mesh;
    einsy_hinge_1: THREE.Mesh;
    einsy_hinge_2: THREE.Mesh;
    frame_linear_bearing_1_1: THREE.Mesh;
    frame_linear_bearing_1_2: THREE.Mesh;
    frame_threaded_shaft_1: THREE.Mesh;
    frame_threaded_shaft_2: THREE.Mesh;
    frame_threaded_shaft_3: THREE.Mesh;
    idler_roller_flat_1: THREE.Mesh;
    idler_roller_flat_2: THREE.Mesh;
    spool_holder: THREE.Mesh;
    x_end_idler_1: THREE.Mesh;
    x_end_motor: THREE.Mesh;
    x_smooth_rod_1: THREE.Mesh;
    x_smooth_rod_2: THREE.Mesh;
    y_belt_idler: THREE.Mesh;
    y_holder_ziptie_1: THREE.Mesh;
    y_holder_ziptie_2: THREE.Mesh;
    y_holder_ziptie_3: THREE.Mesh;
    y_holder_ziptie_4: THREE.Mesh;
    y_motor_holder: THREE.Mesh;
    y_pulley_1: THREE.Mesh;
    y_pulley_2: THREE.Mesh;
    y_rod_holder_1: THREE.Mesh;
    y_rod_holder_2: THREE.Mesh;
    y_rod_holder_3: THREE.Mesh;
    y_rod_holder_4: THREE.Mesh;
    y_smooth_rod_1: THREE.Mesh;
    y_smooth_rod_2: THREE.Mesh;
    z_axis_bottom_left: THREE.Mesh;
    z_axis_bottom_right: THREE.Mesh;
    z_axis_top_left: THREE.Mesh;
    z_axis_top_right: THREE.Mesh;
    z_nut_left: THREE.Mesh;
    z_nut_right: THREE.Mesh;
    z_screw_cover_1: THREE.Mesh;
    z_screw_cover_2: THREE.Mesh;
    z_smooth_rod_1: THREE.Mesh;
    z_smooth_rod_2: THREE.Mesh;
  };
  materials: {
    PaletteMaterial001: THREE.MeshStandardMaterial;
  };
};